# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ICT Helpdesk & Asset Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ticket {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** hardware | software | network | user_support */
  category: string;
  subcategory: string;
  /** low | medium | high | critical */
  priority: string;
  /** open | in_progress | on_hold | resolved | closed */
  status: string;
  reportedBy: string;
  /** @nullable */
  reporterEmail?: string | null;
  /** @nullable */
  reporterDepartment?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  /** @nullable */
  assetId?: number | null;
  /** @nullable */
  resolution?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  resolvedAt?: string | null;
}

export interface TicketInput {
  /** @minLength 3 */
  title: string;
  description?: string;
  category: string;
  subcategory: string;
  priority: string;
  reportedBy: string;
  reporterEmail?: string;
  reporterDepartment?: string;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assetId?: number | null;
}

export interface TicketUpdate {
  title?: string;
  description?: string;
  category?: string;
  subcategory?: string;
  priority?: string;
  status?: string;
  /** @nullable */
  assignedTo?: number | null;
  resolution?: string;
}

export interface TicketComment {
  id: number;
  ticketId: number;
  authorName: string;
  content: string;
  isInternal?: boolean;
  createdAt: string;
}

export interface TicketCommentInput {
  authorName: string;
  content: string;
  isInternal?: boolean;
}

export interface Asset {
  id: number;
  assetTag: string;
  name: string;
  /** desktop | laptop | printer | scanner | server | network | gis | ups | plotter | cctv | other */
  category: string;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  warrantyExpiry?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  /** active | inactive | under_repair | disposed */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AssetInput {
  assetTag?: string;
  name: string;
  category: string;
  manufacturer?: string;
  model?: string;
  serialNumber?: string;
  purchaseDate?: string;
  warrantyExpiry?: string;
  location?: string;
  /** @nullable */
  assignedTo?: number | null;
  status?: string;
  notes?: string;
}

export interface AssetUpdate {
  name?: string;
  category?: string;
  manufacturer?: string;
  model?: string;
  serialNumber?: string;
  purchaseDate?: string;
  warrantyExpiry?: string;
  location?: string;
  /** @nullable */
  assignedTo?: number | null;
  status?: string;
  notes?: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface AssetStats {
  total: number;
  active: number;
  underRepair: number;
  inactive: number;
  disposed: number;
  byCategory: CategoryCount[];
}

export interface InventoryItem {
  id: number;
  name: string;
  /** toner | cable | router | switch | hard_drive | memory | battery | connector | other */
  category: string;
  quantity: number;
  minimumStock: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  notes?: string | null;
  isLowStock?: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface InventoryItemInput {
  name: string;
  category: string;
  quantity: number;
  minimumStock: number;
  unit?: string;
  location?: string;
  supplier?: string;
  notes?: string;
}

export interface InventoryItemUpdate {
  name?: string;
  category?: string;
  quantity?: number;
  minimumStock?: number;
  unit?: string;
  location?: string;
  supplier?: string;
  notes?: string;
}

export interface MaintenanceSchedule {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  assetCategory: string;
  /** @nullable */
  assetId?: number | null;
  /** @nullable */
  assetName?: string | null;
  /** daily | weekly | monthly | quarterly | annually | one_time */
  frequency: string;
  scheduledDate: string;
  /** @nullable */
  completedDate?: string | null;
  /** scheduled | in_progress | completed | overdue | cancelled */
  status: string;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface MaintenanceScheduleInput {
  title: string;
  description?: string;
  assetCategory: string;
  /** @nullable */
  assetId?: number | null;
  frequency: string;
  scheduledDate: string;
  /** @nullable */
  assignedTo?: number | null;
  notes?: string;
}

export interface MaintenanceScheduleUpdate {
  title?: string;
  description?: string;
  assetCategory?: string;
  /** @nullable */
  assetId?: number | null;
  frequency?: string;
  scheduledDate?: string;
  completedDate?: string;
  status?: string;
  /** @nullable */
  assignedTo?: number | null;
  notes?: string;
}

export interface KnowledgeArticle {
  id: number;
  title: string;
  /** troubleshooting | procedure | policy | faq | manual */
  category: string;
  content: string;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  author?: string | null;
  viewCount?: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface KnowledgeArticleInput {
  title: string;
  category: string;
  content: string;
  tags?: string;
  author?: string;
}

export interface KnowledgeArticleUpdate {
  title?: string;
  category?: string;
  content?: string;
  tags?: string;
  author?: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** admin | ict_officer | user */
  role: string;
  department: string;
  /** @nullable */
  phone?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface UserInput {
  name: string;
  email: string;
  role: string;
  department: string;
  phone?: string;
}

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
  department?: string;
  phone?: string;
  isActive?: boolean;
}

export interface DashboardSummary {
  totalTickets: number;
  openTickets: number;
  inProgressTickets: number;
  resolvedToday: number;
  totalAssets: number;
  activeAssets: number;
  assetsUnderRepair: number;
  lowStockItems: number;
  upcomingMaintenance: number;
  overdueMaintenace: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type ListTicketsParams = {
status?: string;
category?: string;
priority?: string;
/**
 * @nullable
 */
assignedTo?: number | null;
};

export type ListAssetsParams = {
category?: string;
status?: string;
location?: string;
};

export type ListInventoryParams = {
lowStock?: boolean;
};

export type ListMaintenanceSchedulesParams = {
status?: string;
upcoming?: boolean;
};

export type ListKnowledgeArticlesParams = {
category?: string;
search?: string;
};