# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ICT Helpdesk & Asset Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all ICT assets
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Register a new ICT asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Maintenance ──────────────────────────────────────────────────────────
...
  /knowledge:
    get:
      operationId: listKnowledgeArticles
      tags: [knowledge]
      summary: List knowledge base articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeArticle"
    post:
      operationId: createKnowledgeArticle
      tags: [knowledge]
      summary: Create knowledge article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeArticleInput"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticle"
  /knowledge/{id}:
    get:
      operationId: getKnowledgeArticle
      tags: [knowledge]
      summary: Get knowledge article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticle"
    patch:
      operationId: updateKnowledgeArticle
      tags: [knowledge]
      summary: Update knowledge article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeArticleUpdate"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticle"
    delete:
      operationId: deleteKnowledgeArticle
      tags: [knowledge]
      summary: Delete knowledge article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Users ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Asset:
      type: object
      required:
        - id
        - assetTag
        - name
        - category
        - status
        - createdAt
      properties:
        id:
          type: integer
        assetTag:
          type: string
        name:
          type: string
        category:
          type: string
          description: "desktop | laptop | printer | scanner | server | network | gis | ups | plotter | cctv | other"
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        purchaseDate:
          type: ["string", "null"]
        warrantyExpiry:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        status:
          type: string
          description: "active | inactive | under_repair | disposed"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    AssetInput:
      type: object
      required:
        - name
        - category
      properties:
        assetTag:
          type: string
        name:
          type: string
        category:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        purchaseDate:
          type: string
        warrantyExpiry:
          type: string
        location:
          type: string
        assignedTo:
          type: ["integer", "null"]
        status:
          type: string
        notes:
          type: string
    AssetUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        purchaseDate:
          type: string
        warrantyExpiry:
          type: string
        location:
          type: string
        assignedTo:
          type: ["integer", "null"]
        status:
          type: string
        notes:
          type: string
...
    InventoryItem:
      type: object
      required:
        - id
        - name
        - category
        - quantity
        - minimumStock
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "toner | cable | router | switch | hard_drive | memory | battery | connector | other"
        quantity:
          type: integer
        minimumStock:
          type: integer
        unit:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isLowStock:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    InventoryItemInput:
      type: object
      required:
        - name
        - category
        - quantity
        - minimumStock
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        minimumStock:
          type: integer
        unit:
          type: string
        location:
          type: string
        supplier:
          type: string
        notes:
          type: string
    InventoryItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        minimumStock:
          type: integer
        unit:
          type: string
        location:
          type: string
        supplier:
          type: string
        notes:
          type: string

    # ── Maintenance schemas ────────────────────────────────────────────────
...
    KnowledgeArticle:
      type: object
      required:
        - id
        - title
        - category
        - content
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          description: "troubleshooting | procedure | policy | faq | manual"
        content:
          type: string
        tags:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        viewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    KnowledgeArticleInput:
      type: object
      required:
        - title
        - category
        - content
      properties:
        title:
          type: string
        category:
          type: string
        content:
          type: string
        tags:
          type: string
        author:
          type: string
    KnowledgeArticleUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        content:
          type: string
        tags:
          type: string
        author:
          type: string

    # ── User schemas ───────────────────────────────────────────────────────
...