# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ICT Helpdesk & Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  assetTag: string;
  name: string;
  /** desktop | laptop | printer | scanner | server | network | gis | ups | plotter | cctv | other */
  category: string;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  warrantyExpiry?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  /** active | inactive | under_repair | disposed */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}