# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ICT Helpdesk & Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  name?: string;
  category?: string;
  quantity?: number;
  minimumStock?: number;
  unit?: string;
  location?: string;
  supplier?: string;
  notes?: string;
}