# lib/api-zod/src/generated/types/knowledgeArticleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ICT Helpdesk & Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export interface KnowledgeArticleUpdate {
  title?: string;
  category?: string;
  content?: string;
  tags?: string;
  author?: string;
}